refactor: reorganize sidebar components for improved structure
- Moved LanguageSwitcher and ThemeToggle to the new sidebar directory for better organization.
- Introduced AppSidebar component to streamline sidebar navigation, integrating user and project links.
- Added NavMain and NavUser components to enhance sidebar functionality and user experience.
- Implemented useIsMobile hook for responsive design, ensuring a seamless experience across devices.
- Updated import paths throughout the codebase to reflect the new structure, enhancing clarity and maintainability.
