+++ src/hooks/useMobile.ts
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 const MOBILE_BREAKPOINT = 768;
 
-export function useIsMobile() {
+export const useIsMobile = () => {
   const [isMobile, setIsMobile] = React.useState<boolean | undefined>(
     undefined
   );
@@ -18,4 +18,4 @@ export function useIsMobile() {
   }, []);
 
   return !!isMobile;
-}
+};
